==> code/1_unir_bussStops.ipynb <==
#!pip install pandas
#!pip install numpy

import pandas as pd
import numpy as np

# Analisis de trayectorio

Dataset de la informacion de paraderos (busstop.csv):
- Muestra informacion de los paraderos que se utilizan durante el trayectorio de los autobuses, tanto como la latitud y longitud.
- Se consideran puntos gps en formato de latitud y longitud


Se procesa la nueva trayectoria de paraderos y analizar los radios que se les asignaran para el geofencing, creando un nuevo archivo de la trayectoria de paraderos.

ida = pd.read_csv('../data/busstops_ida.csv', sep=',')
ida = ida.drop(columns=['radio', 'number'])
ida['tipo'] = ['ida' for _ in range(len(ida))]
#ida

vuelta =  pd.read_csv('../data/busstop.csv', sep=',')
vuelta = vuelta.drop(columns=['number'])
vuelta = vuelta.tail(152-81)
vuelta['tipo'] = ['vuelta' for _ in range(len(vuelta))]
#vuelta

allbusstop_df = pd.concat([ida, vuelta], ignore_index=True)
allbusstop_df['id'] = [f'{i}' for i in range(len(allbusstop_df))]
print(allbusstop_df.info())
allbusstop_df

# Guarda el DataFrame modificado en un nuevo archivo CSV
allbusstop_df.to_csv('../data/allbusstop.csv', index=False)
