==> code/2-analisis de puntos.ipynb <==
import pandas as pd
import numpy as np

# 2. Analisis de los registros de monitorizacion

Tareas a realizar:
- Filtrar los paraderos de vuelta (los que se usa y eliminar los que no se usan y que conincidan con el de ida)

Se da nuevo formato a la tabla, creando nuevas columnas y eliminando otras.

# cargar datos en bruto
monitoringdf = pd.read_csv('../data/_Process_monitoring__202303251137.csv', sep=',')
monitoringdf = monitoringdf[['id', 'latitude','longitude','date','time', 'lap', 'navigation', 'vehicle_id']]
monitoringdf

# Cleaning

Mostramos la cantidad de datos, existen puntos que por error de lectura se leen dos veces, estos puntos se detectan si en las columnas datetime y latitud-longitud existen datos iguales, esto no debe suceder ya que el bus esta en constante movimiento como el tiempo. Como consecuencia se decide eliminar estos duplicados.

# DUPLICADOS
print(f'Shape dataset: {monitoringdf.shape}')

# Verificar y contar si hay puntos duplicados segun datetime y latitude_longitude
base = ['date', 'time', 'latitude', 'longitude']
twocols_dupli = monitoringdf.duplicated(subset=base)
print(f"Nro de duplicados en {base}: {twocols_dupli.sum()}")

# ============================ Solo para comprobacion ============================
# Verificar y contar si hay puntos duplicados segun latitude_longitude
# onecol_dupli =  monitoringdf['latitude_longitude'].duplicated()
# print(len(twocols_dupli), len(onecol_dupli))
# print(twocols_dupli.sum(), onecol_dupli.sum())
# print(abs(twocols_dupli.sum() - onecol_dupli.sum()))
# tmp = [ not x==y for x,y in zip(twocols_dupli, onecol_dupli)]
# data_tmp = monitoringdf.copy()
# data_tmp['dupli'] = tmp
# data_tmp[data_tmp['dupli'] == True]
# ================================================================================

# Eliminar columnas
monitoringdf.drop_duplicates(subset=base, inplace=True)
print(f"Duplicados eliminados => shape dataset: {monitoringdf.shape}")

# NANs
# verificar si hay datos faltantes
print("Cantidad de NaNs:")
for col in monitoringdf:
    nans = monitoringdf[col].isna().sum()
    print(f"\tColumna {col}: {nans}")

# preparar el dataset para iterar en todos los posibles samples
samples_df = monitoringdf.copy()

# Crear columna sample para identificar los samples, utilizando la fecha vehiculo y lapso, todo ello encapsulado en un string
samples_df['sample'] = samples_df.apply(lambda row: f"{row['date']} {row['vehicle_id']} {row['lap']}" , axis=1)
samples_dict = {k:v for v, k in enumerate(samples_df['sample'].unique())}
samples_df['sample'] = samples_df['sample'].apply(lambda sample: samples_dict[sample])
samples_df.sort_values(by='sample', inplace=True)
print('nro posibles samples:', len(samples_dict))

# convertir date and time en el formato datetime
samples_df['datetime'] = samples_df['date'] + ' ' + samples_df['time']
samples_df['datetime'] = pd.to_datetime(samples_df['datetime'], format='%Y-%m-%d %H:%M:%S')
samples_df.drop(columns=['date', 'time'], inplace=True)

samples_df

# Analisis de Samples

# Funcion para extraer un sample por id o nombre y csv para graficar
def getSample(id=-1, name='', to_csv=False):
    if(name != ''):
        id = samples_dict.get(name, -1)
    
    if (id != -1):
        tmp = samples_df[samples_df['sample'] == id].sort_values(by='datetime').reset_index(drop=True)
        print(tmp.shape)
        name = f"{str(tmp.loc[0,'datetime']).split()[0]} {tmp.loc[0,'vehicle_id']} {tmp.loc[0,'lap']}"
        print(tmp.to_string())
        if to_csv: 
            tmp.to_csv(f'{name}- {id} -sample.csv', index=False)
            
    else:
        print(f'No existe una muestra {id} o {name}')

Comprobamos si existen casillas vacias (Nans) en todo el dataframe.

# NANs
# verificar si hay datos faltantes
print("Cantidad de NaNs:")
for col in monitoringdf:
    nans = monitoringdf[col].isna().sum()
    print(f"\tColumna {col}: {nans}")

# preparar el dataset para iterar en todos los posibles samples
samples_df = monitoringdf.copy()

# Crear columna sample para identificar los samples, utilizando la fecha vehiculo y lapso, todo ello encapsulado en un string
samples_df['sample'] = samples_df.apply(lambda row: f"{row['date']} {row['vehicle_id']} {row['lap']}" , axis=1)
samples_dict = {k:v for v, k in enumerate(samples_df['sample'].unique())}
samples_df['sample'] = samples_df['sample'].apply(lambda sample: samples_dict[sample])
samples_df.sort_values(by='sample', inplace=True)
print('nro posibles samples:', len(samples_dict))

# convertir date and time en el formato datetime
samples_df['datetime'] = samples_df['date'] + ' ' + samples_df['time']
samples_df['datetime'] = pd.to_datetime(samples_df['datetime'], format='%Y-%m-%d %H:%M:%S')
samples_df.drop(columns=['date', 'time'], inplace=True)

samples_df

# Funcion para extraer un sample por id o nombre y csv para graficar
def getSample(id=-1, name='', to_csv=False):
    if(name != ''):
        id = samples_dict.get(name, -1)
    
    if (id != -1):
        tmp = samples_df[samples_df['sample'] == id].sort_values(by='datetime').reset_index(drop=True)
        print(tmp.shape)
        name = f"{str(tmp.loc[0,'datetime']).split()[0]} {tmp.loc[0,'vehicle_id']} {tmp.loc[0,'lap']}"
        print(tmp.to_string())
        if to_csv: 
            tmp.to_csv(f'{name}- {id} -sample.csv', index=False)
    else:
        print(f'No existe una muestra {id} o {name}')

# cantidad de laps y vehiculos
veh = samples_df['vehicle_id'].unique()
lap = samples_df['lap'].unique()

print(len(veh))
print(sorted(veh))

print(len(lap))
print(sorted(lap))

# cantidad de puntos por LAP
samples_df.groupby('lap').size()

# Nro de samples por LAP
cp = []
for lap in range(9):
    v = samples_df[samples_df['lap'] == lap]['sample'].unique().shape[0]
    cp.append(v)

# lap = indice de la lista
cp

# id de samples por LAP
samples_df[samples_df['lap'] == 7]['sample'].unique()

samples_sizes = samples_df.groupby('sample').size().to_list()
print('Max:', max(samples_sizes))
print('Min:', min(samples_sizes))

# cantidad de samples por nro de puntos
count, bins = np.histogram(samples_sizes, bins=10, range=(0, 2400))
bins = np.round(bins, 0)
ranges = {f'{int(bins[i])}-{int(bins[i+1])-1}':int(count[i]) for i in range(len(count))}
ranges

lap0_sizes = samples_df[samples_df['lap'] == 0].groupby('sample').size().to_list()
c, b = np.histogram(lap0_sizes, bins=10)
b = np.round(b, 0)
r = {f'{int(b[i])}-{int(b[i+1])-1}':int(c[i]) for i in range(len(c))}
print('max:', max(lap0_sizes))
print('min:', min(lap0_sizes))
r

samples_dfp = samples_df[(samples_df['lap'] != 0) & (samples_df['lap'] != 8) ].copy()
samples_dfp.shape

samples_dfp['lap'].unique()

sizes = samples_dfp.groupby('sample').size()
max(sizes), min(sizes)

# Nro de samples por LAP
ids_mins = []
for l in range(1, 8):
    v = samples_dfp[samples_dfp['lap'] == l]['sample'].unique()
    sizes = samples_dfp[samples_dfp['lap'] == l].groupby('sample').size().to_dict()
    print('lap:' ,l, len(v), max(sizes), min(sizes))
    ids_mins.append(sizes)

samples_dfp.groupby('sample').size().to_dict()

En base al analisis se determina que lap 0 no posee datos que ayuden en la prediccion por lo que seran eliminados

fff = samples_dfp.groupby('sample').size().to_list()
print('max', max(fff))
print('min', min(fff))
ccc, bbb = np.histogram(fff, bins=1, range=(0, 220))
bbb = np.round(bbb, 0)
{f'{int(bbb[i])}-{int(bbb[i+1])-1}':int(ccc[i]) for i in range(len(ccc))}

d = samples_dfp.groupby('sample').size().to_dict()
ids_to_drop = {}
for k, v in d.items():
    if v <= 220:
        ids_to_drop[k]=v
        
len(ids_to_drop), ids_to_drop
