==> code/2_generar_samples.ipynb <==
import pandas as pd

# leer puntos 
monitoringdf = pd.read_csv('../data/_Process_monitoring__202303251137.csv', sep=',')
monitoringdf = monitoringdf[['id', 'latitude','longitude','date','time', 'lap', 'navigation', 'vehicle_id']]
monitoringdf.rename(columns={'id': 'id_point'}, inplace=True)
monitoringdf

# preparar el dataset para iterar en todos los posibles samples
tmp_df = monitoringdf.copy()

# sample_tag = date + vehicle_id + lap : Para identificar un sample
tmp_df['sample_tag'] = tmp_df.apply(lambda row: f"{row['date']} {row['vehicle_id']} {row['lap']}" , axis=1)
samples_dict = {k:v for v, k in enumerate(tmp_df['sample_tag'].unique())}
print('nro posibles samples:', len(samples_dict))

# sample_id = index del sample : Para filtrar por sample
tmp_df['sample_id'] = tmp_df['sample_tag'].apply(lambda sample: samples_dict[sample])
tmp_df.sort_values(by='sample_id', inplace=True)

# date + time = datetime
tmp_df['datetime'] = tmp_df['date'] + ' ' + tmp_df['time']
tmp_df['datetime'] = pd.to_datetime(tmp_df['datetime'], format='%Y-%m-%d %H:%M:%S')
tmp_df.info()

# reordenar columnas
samples_df = tmp_df[['sample_id', 'sample_tag', 'datetime', 'vehicle_id', 'lap', 'latitude', 'longitude', 'navigation', 'id_point']]
samples_df

#samples_df.to_csv('./../data/samples.csv', sep=',', index=False)
