==> code/3_generar_dataset.ipynb <==
import importlib
import pandas as pd
import numpy as np

# Leer datos
samples_df = pd.read_csv('./../data/samples.csv') 
busstop_df = pd.read_csv('./../data/allbusstop.csv')

# damos el formato correcto al campo datetime
samples_df['datetime'] = pd.to_datetime(samples_df['datetime'])

import functions as f; importlib.reload(f)

results = []
for n, points_df in samples_df.groupby(by='sample_id'):
    if n >= 1: break 
    print(f'Evaluando sample_id: {points_df.loc[0, 'sample_id']} con {points_df.shape}' )    
    
    # Para colocar el punto mas cercado a buss stop 
    df_stops = busstop_df.copy()
    df_stops['closest_point'] = df_stops.apply(lambda row_stop: f.closest_point(row_stop, points_df), axis=1)

df_stops
